copmplete rating service, user service

+++ NashShop_BackendApi/Data/Entities/Rating.cs
@@ -10,7 +10,7 @@ namespace NashShop_BackendApi.Data.Entities
         public int Id { get; set; }
         public Guid UserId { get; set; }
         public int ProductId { get; set; }
-        public int Stars { get; set; }
+        public double Stars { get; set; }
         public DateTime DateCreated { get; set; }
         public User User { get; set; }
         public Product Product { get; set; }
